Config changes needed to build on CentOS 6 with newer GCC
Due to a newer-than-standard GCC being installed on our system (6.3.1),
some additional changes (mostly pulled from the Ubuntu 16.04 config) were
needed to get a full build.

The 'url' additions to device-mapper and libvirt were needed to allow
the tarballs to be extracted.

The '--with-user'/'--with-group' additions to device-mapper were needed
to build as non-root (otherwise installation of certain files failed as
it tried to chown them).